Moved all traversal code to seperate file
